Split Dataset-level tests into test_dataset.py and added version/schema_version tests
